Move Panache constants out of PanacheEntityEnhancer
Necessary if we want to split that class in two
