+++ extensions/panache/panache-common/deployment/src/main/java/io/quarkus/panache/common/deployment/visitors/KotlinPanacheClassVisitor.java
@@ -5,7 +5,7 @@ import static io.quarkus.deployment.util.AsmUtil.getLoadOpcode;
 import static io.quarkus.deployment.util.AsmUtil.getSignature;
 import static io.quarkus.deployment.util.AsmUtil.unboxIfRequired;
 import static io.quarkus.gizmo.Gizmo.ASM_API_VERSION;
-import static io.quarkus.panache.common.deployment.PanacheEntityEnhancer.DOTNAME_GENERATE_BRIDGE;
+import static io.quarkus.panache.common.deployment.PanacheConstants.DOTNAME_GENERATE_BRIDGE;
 import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
 import static org.objectweb.asm.Opcodes.ACC_BRIDGE;
